requirements.txt
gspread=6.2.1
python-dotenv=0.21.0
google-auth=2.28.1
google-auth-oauthlib=0.7.1
google-api-python-client=2.101.0
psycopg2-binary=2.9.6
sqlalchemy=2.0.20
sqlalchemy-utils=0.38.4
pandas=2.1.3